get rid of adminController, move show#teams into teamsController
