+++ spec/requests/v1/ssj/dashboard_spec.rb
@@ -54,17 +54,6 @@ RSpec.describe "V1::SSJ::Dashboard", type: :request do
     end
   end
 
-  describe "GET /v1/ssj/dashboard/team" do
-    let(:workflow) { create(:workflow_instance_workflow) }
-
-    it "succeeds" do
-      get "/v1/ssj/dashboard/team", headers: headers
-      expect(response).to have_http_status(:success)
-      expect(json_response["data"]["attributes"]["hasPartner"]).to be false
-      expect(json_response["data"]["attributes"]["expectedStartDate"]).to eq(expected_start_date.to_formatted_s("yyyy-mm-dd"))
-    end
-  end
-
   describe "PUT /v1/ssj/dashboard/team" do
     let(:new_start_date) { "2023-03-01" }
 
@@ -89,7 +78,7 @@ RSpec.describe "V1::SSJ::Dashboard", type: :request do
         }
       }
       expect(response).to have_http_status(:success)
-      expect(json_response["data"]["attributes"]["hasPartner"]).to eq(true)
+      expect(json_response["hasPartner"]).to eq(true)
       person = user.person
       team = person.ssj_team
       expect(SSJ::TeamMember.where(ssj_team_id: team.id, status: SSJ::TeamMember::INVITED)).to_not be_empty
